Add MIT License to the project
The Intelligent Traffic Management System project is released under the MIT License, which is one of the most widely used open-source software licenses. This license allows users to freely use, copy, modify, merge, publish, distribute, and even sell copies of the software, provided that proper credit is given to the original author.